+++ components/tailwind.config.js
@@ -26,7 +26,9 @@ module.exports = {
       bold: 700,
       extrabold: 800,
       black: 900,
-    },
+    }
   },
-  plugins: [],
+  plugins: [
+    require('@tailwindcss/forms'),
+  ],
 };